Fix labels
