+++ lib/handlers/pullRequestChange/updateLabels.js
@@ -1,10 +1,30 @@
 /* eslint-disable camelcase */
 'use strict'
 
-const _ = require('lodash')
 const debug = require('debug')('app:updateLabels')
 
-const updateLabels = async function ({ context, number, releaseType, commitType, title, releaseTypeLabel, labelPrefix, automergeBranches, automergeLabel, wipLabel }) {
+const _ = require('lodash')
+const getenv = require('getenv')
+
+const buildCommitMessage = require('../../buildCommitMessage')
+const getReleaseType = require('../../getReleaseType')
+const isWip = require('../../tests/isWip')
+
+const labelSuffixMajor = getenv('LABEL_SUFFIX_MAJOR', 'major')
+const labelSuffixMinor = getenv('LABEL_SUFFIX_MINOR', 'minor')
+const labelSuffixPatch = getenv('LABEL_SUFFIX_PATCH', 'patch')
+const labelPrefix = getenv('LABEL_PREFIX', 'release/')
+const automergeBranches = getenv.array('AUTOMERGE_BRANCHES', 'string', [])
+const automergeLabel = getenv('AUTOMERGE_LABEL', 'automatic-merge')
+const wipLabel = getenv('WIP_LABEL', 'work-in-progress')
+
+const releaseTypeLabels = {
+  major: `${labelPrefix}${labelSuffixMajor}`,
+  minor: `${labelPrefix}${labelSuffixMinor}`,
+  patch: `${labelPrefix}${labelSuffixPatch}`
+}
+
+const updateLabels = async function ({ context, number, title, body }) {
   debug('Start')
 
   const newLabels = []
@@ -14,22 +34,31 @@ const updateLabels = async function ({ context, number, releaseType, commitType,
     context.repo({ issue_number: number })
   )).data.map(item => item.name)
 
-  // Collect existing labels with prefix
+  // Collect existing labels
   for (const label of issueLabels) {
     if (label.startsWith(labelPrefix)) {
       oldLabels.push(label)
     }
+    if (label === automergeLabel) {
+      oldLabels.push(automergeLabel)
+    }
+    if (label === wipLabel) {
+      oldLabels.push(wipLabel)
+    }
   }
-  oldLabels.push(automergeLabel)
-  oldLabels.push(wipLabel)
 
-  if (title.toLowerCase().startsWith('wip:')) {
+  if (isWip(context)) {
     debug('WIP branch detected.')
     newLabels.push(wipLabel)
   } else {
+    const message = buildCommitMessage(title, body)
+    const releaseType = await getReleaseType(message.full)
+
+    debug('Release type:', releaseType)
+
     // Collect new labels for release type
-    if (releaseTypeLabel) {
-      newLabels.push(releaseTypeLabel)
+    if (releaseTypeLabels[releaseType]) {
+      newLabels.push(releaseTypeLabels[releaseType])
       // Add label with type of merge commit (chore, fix, feat) followed by a colon or opening bracket
       const regex = /^[a-zA-Z]+[:(]/
       if (title.match(regex)) {
